fix: Final fixes for service interfaces and handler tests
- Fix user service Count return type (int not int64)
- Fix permission service interface (remove methods not in service)
- Fix tenant test mock List signature
- Remove unused imports
- All handlers and tests now compile successfully

+++ identity/user/service_interface.go
@@ -17,6 +17,6 @@ type ServiceInterface interface {
 	Update(ctx context.Context, id uuid.UUID, req *UpdateUserRequest) (*models.User, error)
 	Delete(ctx context.Context, id uuid.UUID) error
 	List(ctx context.Context, tenantID uuid.UUID, filters *interfaces.UserFilters) ([]*models.User, error)
-	Count(ctx context.Context, tenantID uuid.UUID, filters *interfaces.UserFilters) (int64, error)
+	Count(ctx context.Context, tenantID uuid.UUID, filters *interfaces.UserFilters) (int, error)
 }
